fix: テスト初期化をbeforeAllから排除し並列実行を安定化
- 各テストのbeforeAll内の動的import(../app)が並列実行時に
  フックタイムアウト(10秒)を超える原因だったため静的importに変更
- Bunシムとフロント配布ディレクトリの準備をvitestのsetupFilesへ集約
- vitest.config.tsとtest-utils/vitest.setup.tsを新規追加
