+++ backend/src/shoppingList/getShoppingList.test.ts
@@ -1,5 +1,4 @@
-import { beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
-import { installBunStaticShim } from '../test-utils/bun-static-shim';
+import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 const { listMenusInRangeMock, findRecipeWithIngredientsMock } = vi.hoisted(() => ({
   listMenusInRangeMock: vi.fn(),
@@ -26,12 +25,7 @@ vi.mock('../shared/auth', () => ({
   getUserId: () => 'user-123',
 }));
 
-let app: (typeof import('../app'))['default'];
-
-beforeAll(async () => {
-  installBunStaticShim();
-  ({ default: app } = await import('../app'));
-});
+import app from '../app';
 
 const getShoppingListRequest = (from: string, to: string): Promise<Response> =>
   app.request(`/api/shopping-list?from=${from}&to=${to}`);
